+++ src/main/java/pl/dicedev/kakebo/security/UserDetailsRepository.java
@@ -2,7 +2,7 @@ package pl.dicedev.kakebo.security;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-import pl.dicedev.kakebo.security.entity.UserEntity;
+import pl.dicedev.kakebo.repositories.entities.UserEntity;
 
 import java.util.Optional;
 import java.util.UUID;
