+++ src/test/java/pl/dicedev/kakebo/security/UserServiceImplTest.java
@@ -7,8 +7,8 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
+import pl.dicedev.kakebo.repositories.entities.UserEntity;
 import pl.dicedev.kakebo.security.dto.AuthUserDto;
-import pl.dicedev.kakebo.security.entity.UserEntity;
 import pl.dicedev.kakebo.security.exceptions.UserAlreadyExistException;
 import pl.dicedev.kakebo.security.mapper.UserMapper;
 import pl.dicedev.kakebo.security.mapper.UserMapperImpl;
